Remove weird, unused and breaking behavior
There was an if statement in the parseData function (in the npm module, it is in modules/parse.fn.js) which would return the full Array returned by rgx.exec(data), instead of the String indicating the match itself. The if statement did not seem to be actually used anywhere; it only broke things (for example, inline parsing in inline Array selectors; for example, {"myInfo":["a @ href || [a-z]+"]} would cause an error because parseData (in this case, which had multiple set to true) returned an Array, not a String). This removes that if statement.